@@ -1,10 +1,9 @@
 package co.yovany.androidtestproject.adapter;
 
 import android.app.Activity;
-import android.support.annotation.NonNull;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v7.widget.RecyclerView;
-import android.text.Layout;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -19,7 +18,6 @@ import java.util.ArrayList;
 
 import co.yovany.androidtestproject.R;
 import co.yovany.androidtestproject.model.Student;
-import de.hdodenhof.circleimageview.CircleImageView;
 
 public class StudentAdapterRecyclerView extends RecyclerView.Adapter<StudentAdapterRecyclerView.StudentViewHolder> {
 
