@@ -1,6 +1,6 @@
 package co.yovany.androidtestproject.view;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import co.yovany.androidtestproject.R;
