@@ -1,9 +1,8 @@
 package co.yovany.androidtestproject.view;
 
-import android.app.FragmentManager;
-import android.app.FragmentTransaction;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import co.yovany.androidtestproject.R;
@@ -34,7 +33,7 @@ public class NestedScrollViewActivity extends AppCompatActivity {
     private void launchFragment(Fragment fragment, int resource) {
         getSupportFragmentManager().beginTransaction()
                 .replace(resource, fragment)
-                .setTransition(android.support.v4.app.FragmentTransaction.TRANSIT_FRAGMENT_FADE)
+                .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
                 .commit();
     }
 }
